[PATCH] 'version': vutil.c: silence overflow warnings
From a3f461a88988cef9a89c13ab42c50556750ede51 Mon Sep 17 00:00:00 2001
Running 'clang -fsanitize=undefined' on the perl test suite triggers
a number of warnings:

    vutil.c:354:9: runtime error: signed integer overflow: 750283776 + 1705032704 cannot be represented in type 'int'
    vutil.c:354:25: runtime error: signed integer overflow: 6 * 1000000000 cannot be represented in type 'int'
    vutil.c:355:10: runtime error: signed integer overflow: 1000000000 * 10 cannot be represented in type 'int'

This is because the version string parsing code takes a "maybe overflow,
then check to see if we've overflown" approach. Which is perfectly fine,
but clashes with clang.

This commit makes it take the opposite approach of checking each time
*before* doing the += mult*i and mult*=10 steps.

A bit clunky, but ASan only allows you to disable these warnings on a
per-function basis, and it's a reasonably large function.

I suspect that the block of code above this may need similar treatment,
but as it wasn't triggering clang warnings, I didn't look at it.
---
 vutil.c | 17 ++++++++++++-----
 1 file changed, 12 insertions(+), 5 deletions(-)