Add basic GitHub CI workflow
This is testing the distribution on most
Perl version since 5.10 to 5.32

Note that tests are failing on 5.6 and 5.8 for now.
This would need investigation before enabling them.

A pre-check is run using the basic ubuntu-latest container
with the system perl. Only on success it's going to
trigger the additional jobs.