vutil.c: Use improved algorithm for non-P2008 radix
Prior to this commit, the code went through lots of machinations to
detect the radix character of an NV in order to handle locales in which
the radix is not a dot.  This is because the version is always expressed
in dot notation, and we had to accommodate sprintf(), which uses the
current locale, and that locale might cause the value to be printed with
something other than a dot.

What was done was to change locales to something known to be a dot
locale, so that we would be guaranteed to get a dot.  But on
non-POSIX-2008 boxes, this has significant drawbacks.  On threaded or
embedded perls, changing the locale can clash with other instances on
platforms that don't have thread-safe locale handling.  This can be
overcome by using mutexes, as long as the other instances also obey
those mutexes, and the perl this is compiled with is of sufficiently
modern vintage to have those.

A better method is to not change locale, but take the output of the
sprintf and change the radix to a dot from whatever it got formatted as.
This is feasible because almost all locales in the world have either a
dot or a comma radix.  So, the code now examines the sprintf output and
converts a found comma into a dot.  No need to change locales.

I am aware of only one other potential radix character, found only in a
Pashtun language locale of the 500 locales available on my box.  This
commit also handles that possibility, again converting it to a dot.

Should there be a locale where this doesn't work, the commit leaves the
current state as a fallback, subject to its disadvantages.  But the
chances of this fallback actually getting executed become close to zero.

The POSIX 2008 case from before is both trivial and safe, so is left
as-is.